backend/requirements.txt
Babel==2.9.1
fastapi==0.109.0
frozenlist==1.3.3
magic-filter==1.0.9
Pillow==10.2.0
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.0
starlette==0.36.1
typing_extensions==4.4.0
uvicorn==0.20.0
yarl==1.8.2
aiofiles==23.2.1
